Fix build error for whole CK package build on gfx1300 + windows (#215)
1. Add ROCM_SEARCH_PATH to support windows build
2. Add gfx13 for SUPPORTED_GPU_TARGETS, CK_GPU_TARGETS
3. Add ENABLE_GFX13_FAIL_TESTS to disable unsupported cases
- Modify code to fix build errors

build is passed with following command:
cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DCMAKE_C_COMPILER=C:/opt/rocm/bin/clang.exe -DCMAKE_CXX_COMPILER=C:/opt/rocm/bin/clang++.exe -DGPU_TARGETS="gfx1301" -DGPU_ARCH="gfx13" -DCMAKE_OBJECT_PATH_MAX=512 ..

+++ projects/composablekernel/include/ck/tensor_operation/gpu/element/element_wise_operation.hpp
@@ -411,7 +411,7 @@ struct MultiplyAddTanh
     {
         const float x = c * d0 + d1;
 
-        e = std::fmin((float)1.0, std::fmax((float)-1.0, x));
+        e = std::fmin(1.0f, std::fmax(-1.0f, x));
     }
 
     template <>
@@ -420,7 +420,7 @@ struct MultiplyAddTanh
     {
         const float x = type_convert<float>(c) * type_convert<float>(d0) + type_convert<float>(d1);
 
-        e = type_convert<half_t>(std::fmin((float)1.0, std::fmax((float)-1.0, x)));
+        e = type_convert<half_t>(std::fmin(1.0f, std::fmax(-1.0f, x)));
     }
 
     template <>
@@ -429,7 +429,7 @@ struct MultiplyAddTanh
     {
         const float x = type_convert<float>(c) * type_convert<float>(d0) + type_convert<float>(d1);
 
-        e = type_convert<bhalf_t>(std::fmin((float)1.0, std::fmax((float)-1.0, x)));
+        e = type_convert<bhalf_t>(std::fmin(1.0f, std::fmax(-1.0f, x)));
     }
 
     template <>
@@ -438,7 +438,7 @@ struct MultiplyAddTanh
     {
         const half_t x = c * type_convert<half_t>(d0) + type_convert<half_t>(d1);
 
-        e = type_convert<half_t>(std::fmin((float)1.0, std::fmax((float)-1.0, x)));
+        e = type_convert<half_t>(std::fmin(1.0f, std::fmax(-1.0f, x)));
     }
 };
