+++ projects/composablekernel/test/scatter_gather/scatter_gather.cpp
@@ -144,8 +144,8 @@ union pixel
 {
     struct __attribute__((packed))
     {
-        unsigned int r : 6;
-        unsigned int c : 10;
+        ushort r : 6;
+        ushort c : 10;
     };
     ushort data;
 };
